+++ code/src/main/java/fr/inria/kgindex/rules/RuleApplication.java
@@ -1,16 +1,20 @@
 package fr.inria.kgindex.rules;
 
-import fr.inria.kgindex.data.Dataset;
-import fr.inria.kgindex.data.ManifestEntry;
-import fr.inria.kgindex.data.RuleLibrary;
+import fr.inria.kgindex.data.*;
 import fr.inria.kgindex.util.EarlReport;
+import fr.inria.kgindex.util.KGIndex;
 import fr.inria.kgindex.util.Utils;
 import org.apache.jena.query.QueryExecution;
 import org.apache.jena.query.QueryExecutionFactory;
 import org.apache.jena.query.QueryParseException;
-import org.apache.jena.rdf.model.*;
+import org.apache.jena.query.ResultSet;
+import org.apache.jena.rdf.model.Literal;
+import org.apache.jena.rdf.model.Model;
+import org.apache.jena.rdf.model.ModelFactory;
+import org.apache.jena.rdf.model.RDFNode;
 import org.apache.jena.riot.RiotException;
 import org.apache.jena.sparql.engine.http.QueryExceptionHTTP;
+import org.apache.jena.sparql.resultset.RDFInput;
 import org.apache.jena.sparql.vocabulary.EARL;
 import org.apache.jena.update.UpdateAction;
 import org.apache.jena.update.UpdateFactory;
@@ -18,7 +22,17 @@ import org.apache.jena.update.UpdateRequest;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import java.util.*;
+import java.io.StringReader;
+import java.io.UnsupportedEncodingException;
+import java.net.URI;
+import java.net.URLEncoder;
+import java.net.http.HttpClient;
+import java.net.http.HttpResponse;
+import java.util.ArrayList;
+import java.util.Date;
+import java.util.List;
+import java.util.Set;
+import java.net.http.HttpRequest;
 
 import static fr.inria.kgindex.util.Utils.dateFormatter;
 
@@ -31,6 +45,8 @@ public class RuleApplication {
         SHACL,
         UNKNOWN
     }
+
+    public static String federationserver = null;
 
     private final ManifestEntry _entry;
     private final Dataset _describedDataset;
@@ -94,39 +110,72 @@ public class RuleApplication {
                 String queryStringRaw = action.getActionNode().asLiteral().getString();
                 Set<String> queryStringSet = Utils.rewriteQueryPlaceholders(queryStringRaw, this._describedDataset);
                 queryStringSet.forEach(queryString -> {
-                    Date startDate = new Date();
-                    Literal startDateLiteral = result.createLiteral(dateFormatter.format(startDate));
-                    try {
-                        if (queryString.contains("CONSTRUCT")) {
-                            QueryExecution actionExecution = QueryExecutionFactory.sparqlService(action.getEndpointUrl(), queryString);
-                            actionExecution.setTimeout(Utils.queryTimeout);
-
-                            try {
-                                Model actionResult = actionExecution.execConstruct();
-                                result.add(actionResult);
-                            } catch (RiotException e) {
-                                logger.error(e);
-                                logger.trace(this._entry.getTestResource() + " action could not be added because of RiotException");
-                            }
-                            actionExecution.close();
-                        } else if (queryString.contains("INSERT")) {
-                            UpdateRequest insertUpdate = UpdateFactory.create(queryString);
-                            UpdateAction.execute(insertUpdate, this._datasetDescription);
-                        } else if (queryString.contains("DELETE")) {
-                            UpdateRequest deleteUpdate = UpdateFactory.create(queryString);
-                            UpdateAction.execute(deleteUpdate, this._datasetDescription);
+                    if ((action.getEndpointUrl().equals(KGIndex.federation.getURI()) && (RuleApplication.federationserver != null))
+                            || (!action.getEndpointUrl().equals(KGIndex.federation.getURI()))) {
+                        if (action.getEndpointUrl().equals(KGIndex.federation.getURI())) {
+                            action.setEndpointUrl(RuleApplication.federationserver);
                         }
-                    } catch (QueryExceptionHTTP e) {
-                        logger.info(e);
-                        logger.trace(this._entry.getTestResource() + " : " + e.getMessage());
-                        Date endDate = new Date();
-                        Literal endDateLiteral = result.createLiteral(dateFormatter.format(endDate));
-                        result.add(EarlReport.createEarlFailedQueryReport(this._describedDataset, queryString, this._entry, e.getMessage(), startDateLiteral, endDateLiteral));
-                    } catch (QueryParseException e) {
-                        logger.debug(queryString);
-                        throw e;
-                    }
+                        Date startDate = new Date();
+                        Literal startDateLiteral = result.createLiteral(dateFormatter.format(startDate));
+                        try {
+                            if (queryString.contains("CONSTRUCT")) {
+                                QueryExecution actionExecution = QueryExecutionFactory.sparqlService(action.getEndpointUrl(), queryString);
+                                actionExecution.setTimeout(Utils.queryTimeout);
 
+                                try {
+                                    Model actionResult = actionExecution.execConstruct();
+                                    result.add(actionResult);
+                                } catch (RiotException e) {
+                                    logger.error(e);
+                                    logger.trace(this._entry.getTestResource() + " action could not be added because of RiotException");
+                                }
+                                actionExecution.close();
+                            } else if (queryString.contains("INSERT")) {
+                                UpdateRequest insertUpdate = UpdateFactory.create(queryString);
+                                UpdateAction.execute(insertUpdate, this._datasetDescription);
+                            } else if (queryString.contains("DELETE")) {
+                                UpdateRequest deleteUpdate = UpdateFactory.create(queryString);
+                                UpdateAction.execute(deleteUpdate, this._datasetDescription);
+                            }
+                        } catch (QueryExceptionHTTP e) {
+                            logger.info(e);
+                            logger.trace(this._entry.getTestResource() + " : " + e.getMessage());
+                            Date endDate = new Date();
+                            Literal endDateLiteral = result.createLiteral(dateFormatter.format(endDate));
+                            result.add(EarlReport.createEarlFailedQueryReport(this._describedDataset, queryString, this._entry, e.getMessage(), startDateLiteral, endDateLiteral));
+                        } catch (QueryParseException e) {
+                            // Tentative d'envoyer la requête sans passer par Jena
+                            HttpClient client = HttpClient.newHttpClient();
+                            HttpRequest request = null;
+                            try {
+                                URI queryURL = URI.create(action.getEndpointUrl() + "?query=" + URLEncoder.encode(queryString, java.nio.charset.StandardCharsets.UTF_8.toString()));
+                                logger.debug(queryURL);
+                                request = HttpRequest.newBuilder()
+                                        .uri(queryURL)
+                                        .GET()
+                                        .header("Accept", "application/rdf+xml")
+                                        .build();
+                            } catch (UnsupportedEncodingException e1) {
+                                e1.printStackTrace();
+                            }
+                            client.sendAsync(request, HttpResponse.BodyHandlers.ofString())
+                                    .thenApply(HttpResponse::body)
+                                    .thenAccept(bodyString -> {
+                                        if (queryString.contains("CONSTRUCT")) {
+                                            Model bodyModel = ModelFactory.createDefaultModel();
+                                            StringReader bodyReader = new StringReader(bodyString);
+                                            try {
+                                                bodyModel.read(bodyReader, "");
+                                                result.add(bodyModel);
+                                            } catch(RiotException er) {
+                                                logger.error(bodyString);
+                                                throw e;
+                                            }
+                                        }
+                                    })
+                                    .join();
+                        }
+                    }
                 });
             } else if(action.getType() == Action.TYPE.Manifest) {
                 Set<ManifestEntry> entrySet = RuleLibrary.getLibrary().get(action.getActionNode());