+++ code/src/main/java/fr/inria/kgindex/util/KGIndex.java
@@ -8,9 +8,10 @@ import org.apache.jena.rdf.model.Resource;
 public class KGIndex {
 	
 	private static final Model model = ModelFactory.createDefaultModel();
-	public static final String kgindexNamespace = "http://ns.inria.fr/kgindex#";
+	public static final String kgindexNamespace = "http://ns.inria.fr/kg/index#";
 
 	public static final Resource TestQuery = model.createResource(kgindexNamespace + "TestQuery");
+	public static final Resource federation = model.createResource(kgindexNamespace + "federation");
 	public static final Property query = model.createProperty(kgindexNamespace + "query");
 	public static final Property sentQuery = model.createProperty(kgindexNamespace + "sentQuery");
 	public static final Property trace = model.createProperty(kgindexNamespace + "trace");